pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moelmer-model"
version = "0.1.0"
description = "Config-driven reproduction tools for the Moelmer superradiant lasing model"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "PyYAML>=6.0",
  "scipy>=1.11",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
